Refactor of Commands in Command interpreter
Refactored Commands in Command interprerter
Also some minor bug fixes
