+++ Coffe_Maker_MKI/Debug/services/CommandInterpreter/CommandInterpreter.d
@@ -25,7 +25,26 @@ services/CommandInterpreter/CommandInterpreter.d \
  ../services/CommandInterpreter/../powermeter/powermeter.h \
  c:\program\ files\ (x86)\atmel\studio\7.0\toolchain\avr8\avr8-gnu-toolchain\avr\include\stdlib.h \
  ../services/CommandInterpreter/../boilertempcontroller/boilertempcontroller.h \
- ../services/CommandInterpreter/../../driver/gpio/gpio_typedef.h
+ ../services/CommandInterpreter/CommandInterpreterTypes.h \
+ ../services/CommandInterpreter/CommandInterpreterInternalVars.h \
+ ../services/CommandInterpreter/Commands/CMDI_CHPRG.h \
+ ../services/CommandInterpreter/Commands/../CommandInterpreterTypes.h \
+ ../services/CommandInterpreter/Commands/CMDI_PREWATER.h \
+ ../services/CommandInterpreter/Commands/CMDI_PRG.h \
+ ../services/CommandInterpreter/Commands/CMDI_SYS_HELP.h \
+ ../services/CommandInterpreter/Commands/CMDI_TEMP.h \
+ ../services/CommandInterpreter/Commands/CMDI_WAIT.h \
+ ../services/CommandInterpreter/Commands/CMDI_WATER.h \
+ ../services/CommandInterpreter/Commands/CMDI_MES_POWER.h \
+ ../services/CommandInterpreter/Commands/CMDI_CUP_LEVEL.h \
+ ../services/CommandInterpreter/Commands/CMDI_AUTOPOWER.h \
+ ../services/CommandInterpreter/Commands/CMDI_STBY_TEMP.h \
+ ../services/CommandInterpreter/Commands/CMDI_POWER.h \
+ ../services/CommandInterpreter/Commands/CMDI_ECHO.h \
+ ../services/CommandInterpreter/Commands/CMDI_GPIO.h \
+ ../services/CommandInterpreter/Commands/CMDI_BOILERTEMP.h \
+ ../services/CommandInterpreter/Commands/CMDI_MAKE.h \
+ ../services/CommandInterpreter/Commands/COMMON_VALUE_STORE.h
 
 c:\program\ files\ (x86)\atmel\studio\7.0\toolchain\avr8\avr8-gnu-toolchain\avr\include\string.h:
 
@@ -75,4 +94,42 @@ c:\program\ files\ (x86)\atmel\studio\7.0\toolchain\avr8\avr8-gnu-toolchain\avr\
 
 ../services/CommandInterpreter/../boilertempcontroller/boilertempcontroller.h:
 
-../services/CommandInterpreter/../../driver/gpio/gpio_typedef.h:
+../services/CommandInterpreter/CommandInterpreterTypes.h:
+
+../services/CommandInterpreter/CommandInterpreterInternalVars.h:
+
+../services/CommandInterpreter/Commands/CMDI_CHPRG.h:
+
+../services/CommandInterpreter/Commands/../CommandInterpreterTypes.h:
+
+../services/CommandInterpreter/Commands/CMDI_PREWATER.h:
+
+../services/CommandInterpreter/Commands/CMDI_PRG.h:
+
+../services/CommandInterpreter/Commands/CMDI_SYS_HELP.h:
+
+../services/CommandInterpreter/Commands/CMDI_TEMP.h:
+
+../services/CommandInterpreter/Commands/CMDI_WAIT.h:
+
+../services/CommandInterpreter/Commands/CMDI_WATER.h:
+
+../services/CommandInterpreter/Commands/CMDI_MES_POWER.h:
+
+../services/CommandInterpreter/Commands/CMDI_CUP_LEVEL.h:
+
+../services/CommandInterpreter/Commands/CMDI_AUTOPOWER.h:
+
+../services/CommandInterpreter/Commands/CMDI_STBY_TEMP.h:
+
+../services/CommandInterpreter/Commands/CMDI_POWER.h:
+
+../services/CommandInterpreter/Commands/CMDI_ECHO.h:
+
+../services/CommandInterpreter/Commands/CMDI_GPIO.h:
+
+../services/CommandInterpreter/Commands/CMDI_BOILERTEMP.h:
+
+../services/CommandInterpreter/Commands/CMDI_MAKE.h:
+
+../services/CommandInterpreter/Commands/COMMON_VALUE_STORE.h:
