+++ Coffe_Maker_MKI/services/CommandInterpreter/CommandInterpreterInternalVars.h
@@ -1,0 +1,15 @@
+/*
+ * CommandInterpreterInternalFunctions.h
+ *
+ * Created: 16.09.2023 12:01:40
+ *  Author: CalM
+ */ 
+
+
+#ifndef COMMANDINTERPRETERINTERNALFUNCTIONS_H_
+#define COMMANDINTERPRETERINTERNALFUNCTIONS_H_
+
+volatile CmdConfigBit_t CmdConfig;
+
+
+#endif /* COMMANDINTERPRETERINTERNALFUNCTIONS_H_ */
