+++ Coffe_Maker_MKI/services/CommandInterpreter/Commands/CMDI_AUTOPOWER.h
@@ -1,0 +1,18 @@
+/*
+ * CMDI_AUTOPOWER.h
+ *
+ * Created: 15.09.2023 19:11:51
+ *  Author: CalM
+ */ 
+
+
+#ifndef CMDI_AUTOPOWER_H_
+#define CMDI_AUTOPOWER_H_
+
+#include "../CommandInterpreterTypes.h"
+
+void voCMDI_AUTOPOWER(CMDI_CMD_DATA_t COMMAND);
+
+
+
+#endif /* CMDI_AUTOPOWER_H_ */
