+++ Coffe_Maker_MKI/services/CommandInterpreter/Commands/CMDI_BOILERTEMP.h
@@ -1,0 +1,18 @@
+/*
+ * CMDI_BOILERTEMP.h
+ *
+ * Created: 15.09.2023 18:50:56
+ *  Author: CalM
+ */ 
+
+
+#ifndef CMDI_BOILERTEMP_H_
+#define CMDI_BOILERTEMP_H_
+
+#include "../CommandInterpreterTypes.h"
+
+void voCMDI_BOILERTEMP(CMDI_CMD_DATA_t COMMAND);
+
+
+
+#endif /* CMDI_BOILERTEMP_H_ */
