+++ Coffe_Maker_MKI/services/CommandInterpreter/Commands/CMDI_CHPRG.h
@@ -1,0 +1,17 @@
+/*
+ * CMDI_CHPRG.h
+ *
+ * Created: 15.09.2023 18:43:17
+ *  Author: CalM
+ */ 
+
+
+#ifndef CMDI_CHPRG_H_
+#define CMDI_CHPRG_H_
+
+#include "../CommandInterpreterTypes.h"
+
+void voCMDI_CHPRG(CMDI_CMD_DATA_t COMMAND);
+
+
+#endif /* CMDI_CHPRG_H_ */
