+++ Coffe_Maker_MKI/services/CommandInterpreter/Commands/CMDI_CUP_LEVEL.h
@@ -1,0 +1,14 @@
+/*
+ * CMDI_CUP_LEVEL.h
+ *
+ * Created: 15.09.2023 19:01:31
+ *  Author: CalM
+ */ 
+#ifndef CMDI_CUP_LEVEL_H_
+#define CMDI_CUP_LEVEL_H_
+
+#include "../CommandInterpreterTypes.h"
+
+void voCMDI_CUP_LEVEL(CMDI_CMD_DATA_t COMMAND);
+
+#endif
