+++ Coffe_Maker_MKI/services/CommandInterpreter/Commands/CMDI_ECHO.h
@@ -1,0 +1,19 @@
+/*
+ * CMDI_ECHO.h
+ *
+ * Created: 15.09.2023 19:20:07
+ *  Author: CalM
+ */ 
+
+
+#ifndef CMDI_ECHO_H_
+#define CMDI_ECHO_H_
+
+#include "../CommandInterpreterTypes.h"
+
+void voCMDI_ECHO(CMDI_CMD_DATA_t COMMAND);
+
+
+
+
+#endif /* CMDI_ECHO_H_ */
