+++ Coffe_Maker_MKI/services/CommandInterpreter/Commands/CMDI_GPIO.h
@@ -1,0 +1,18 @@
+/*
+ * CMDI_GPIO.h
+ *
+ * Created: 16.09.2023 11:28:14
+ *  Author: CalM
+ */ 
+
+
+#ifndef CMDI_GPIO_H_
+#define CMDI_GPIO_H_
+
+#include "../CommandInterpreterTypes.h"
+
+void voCMDI_GPIO(CMDI_CMD_DATA_t COMMAND);
+
+
+
+#endif /* CMDI_GPIO_H_ */
