+++ Coffe_Maker_MKI/services/CommandInterpreter/Commands/CMDI_MAKE.h
@@ -1,0 +1,18 @@
+/*
+ * CMDI_MAKE.h
+ *
+ * Created: 16.09.2023 11:31:57
+ *  Author: CalM
+ */ 
+
+
+#ifndef CMDI_MAKE_H_
+#define CMDI_MAKE_H_
+
+#include "../CommandInterpreterTypes.h"
+
+void voCMDI_MAKE(CMDI_CMD_DATA_t COMMAND);
+
+
+
+#endif /* CMDI_MAKE_H_ */
