+++ Coffe_Maker_MKI/services/CommandInterpreter/Commands/CMDI_MES_POWER.h
@@ -1,0 +1,19 @@
+/*
+ * CMDI_MES_POWER.h
+ *
+ * Created: 15.09.2023 18:47:38
+ *  Author: CalM
+ */ 
+
+
+#ifndef CMDI_MES_POWER_H_
+#define CMDI_MES_POWER_H_
+
+#include "../CommandInterpreterTypes.h"
+
+void voCMDI_MES_POWER(CMDI_CMD_DATA_t COMMAND);
+
+
+
+
+#endif /* CMDI_MES_POWER_H_ */
