+++ Coffe_Maker_MKI/services/CommandInterpreter/Commands/CMDI_POWER.h
@@ -1,0 +1,19 @@
+/*
+ * CMDI_POWER.h
+ *
+ * Created: 15.09.2023 19:07:02
+ *  Author: CalM
+ */ 
+
+
+#ifndef CMDI_POWER_H_
+#define CMDI_POWER_H_
+
+#include "../CommandInterpreterTypes.h"
+
+void voCMDI_POWER(CMDI_CMD_DATA_t COMMAND);
+
+
+
+
+#endif /* CMDI_POWER_H_ */
