+++ Coffe_Maker_MKI/services/CommandInterpreter/Commands/CMDI_PREWATER.h
@@ -1,0 +1,14 @@
+/*
+ * CMDI_PREWATER.h
+ *
+ * Created: 13.09.2023 21:02:16
+ *  Author: CalM
+ */ 
+
+#ifndef CMD_PREWATER_H_
+#define CMD_PREWATER_H_
+#include "../CommandInterpreterTypes.h"
+
+void voCMDI_PREWATER(CMDI_CMD_DATA_t COMMAND);
+
+#endif
