+++ Coffe_Maker_MKI/services/CommandInterpreter/Commands/CMDI_PRG.h
@@ -1,0 +1,18 @@
+/*
+ * CMD_Temp.h
+ *
+ * Created: 13.09.2023 20:26:43
+ *  Author: CalM
+ */ 
+
+
+#ifndef CMD_PRG_H_
+#define CMD_PRG_H_
+
+#include "../CommandInterpreterTypes.h"
+
+void voCMDI_PRG( CMDI_CMD_DATA_t COMMAND );
+
+
+
+#endif /* CMD_TEMP_H_ */
