+++ Coffe_Maker_MKI/services/CommandInterpreter/Commands/CMDI_STBY_TEMP.h
@@ -1,0 +1,18 @@
+/*
+ * CMDI_STBY_TEMP.h
+ *
+ * Created: 15.09.2023 19:15:48
+ *  Author: CalM
+ */ 
+
+
+#ifndef CMDI_STBY_TEMP_H_
+#define CMDI_STBY_TEMP_H_
+
+#include "../CommandInterpreterTypes.h"
+
+void voCMDI_STBY_TEMP(CMDI_CMD_DATA_t COMMAND);
+
+
+
+#endif /* CMDI_STBY_TEMP_H_ */
