+++ Coffe_Maker_MKI/services/CommandInterpreter/Commands/CMDI_SYS_HELP.h
@@ -1,0 +1,22 @@
+/*
+ * CMD_SYS_HELP.h
+ *
+ * Created: 13.09.2023 20:29:01
+ *  Author: CalM
+ */ 
+
+
+#ifndef CMD_SYS_HELP_H_
+#define CMD_SYS_HELP_H_
+
+#include "../CommandInterpreterTypes.h"
+/*
+Function	:	voCMDI_SYS_HELP
+Parameter	:	CMDI_CMD_DATA_t
+Returnvalue	:	None
+Remkark		:	Prints Help message and also handles BEER command
+*/
+void voCMDI_SYS_HELP( CMDI_CMD_DATA_t COMMAND );
+
+
+#endif /* CMD_SYS_HELP_H_ */
