+++ Coffe_Maker_MKI/services/CommandInterpreter/Commands/CMDI_WAIT.h
@@ -1,0 +1,14 @@
+/*
+ * CMDI_WAIT.h
+ *
+ * Created: 14.09.2023 19:36:34
+ *  Author: CalM
+ */ 
+#ifndef CMD_WAIT_H_
+#define CMD_WAIT_H_
+
+#include "../CommandInterpreterTypes.h"
+
+void voCMDI_WAIT( CMDI_CMD_DATA_t COMMAND );
+
+#endif
