+++ Coffe_Maker_MKI/services/CommandInterpreter/Commands/CMDI_WATER.h
@@ -1,0 +1,18 @@
+/*
+ * CMDI_WATER.h
+ *
+ * Created: 13.09.2023 20:59:44
+ *  Author: CalM
+ */ 
+
+
+#ifndef CMDI_WATER_H_
+#define CMDI_WATER_H_
+#include "../CommandInterpreterTypes.h"
+
+void voCMDI_WATRER(CMDI_CMD_DATA_t COMMAND);
+
+
+
+
+#endif /* CMDI_WATER_H_ */
