+++ Coffe_Maker_MKI/services/CommandInterpreter/Commands/COMMON_VALUE_STORE.h
@@ -1,0 +1,18 @@
+/*
+ * COMMON_VALUE_STORE.h
+ *
+ * Created: 16.09.2023 11:46:28
+ *  Author: CalM
+ */ 
+
+
+#ifndef COMMON_VALUE_STORE_H_
+#define COMMON_VALUE_STORE_H_
+
+#include "stdint.h"
+
+uint8_t u8GetProgToChange(void);
+void voSetProgToChange(uint8_t value);
+
+
+#endif /* COMMON_VALUE_STORE_H_ */
