+++ Coffe_Maker_MKI/services/CommandInterpreter/Commands/PARSE_RESPONSE.h
@@ -1,0 +1,18 @@
+/*
+ * PARSE_ERROR.h
+ *
+ * Created: 13.09.2023 20:40:08
+ *  Author: CalM
+ */ 
+
+
+#ifndef PARSE_ERROR_H_
+#define PARSE_ERROR_H_
+
+void voResponseParamErr( void );
+void voResponseParserErr( void );
+void voResponseOk( void );
+
+
+
+#endif /* PARSE_ERROR_H_ */
