+++ Coffe_Maker_MKI/services/usb_cdc_helper/usb_cdc_helper.h
@@ -1,0 +1,20 @@
+/*
+ * usb_cdc_helper.h
+ *
+ * Created: 16.09.2023 12:21:02
+ *  Author: CalM
+ */ 
+
+
+#ifndef USB_CDC_HELPER_H_
+#define USB_CDC_HELPER_H_
+
+#include "stdbool.h"
+#include "stdio.h"
+
+bool boGetConnected( void );
+void voSetConnected(bool con);
+FILE* GetUSBStream( void );
+void voSetUSBStream(FILE Stream);
+
+#endif /* USB_CDC_HELPER_H_ */
